[webgl] merge CumProdProgram and CumSumProgram
