+++ tfjs-backend-webgl/src/kernels/Cumsum.ts
@@ -1,6 +1,6 @@
 /**
  * @license
- * Copyright 2020 Google LLC. All Rights Reserved.
+ * Copyright 2022 Google LLC. All Rights Reserved.
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
@@ -15,13 +15,11 @@
  * =============================================================================
  */
 
-import {backend_util, Cumsum, CumsumAttrs, CumsumInputs, KernelConfig, KernelFunc, TensorInfo} from '@tensorflow/tfjs-core';
+import {Cumsum, CumsumAttrs, CumsumInputs, KernelConfig, KernelFunc, TensorInfo} from '@tensorflow/tfjs-core';
 
 import {MathBackendWebGL} from '../backend_webgl';
-import {CumSumProgram} from '../cumsum_gpu';
-
-import {identity} from './Identity';
-import {transpose} from './Transpose';
+import {CumOpType} from '../cum_gpu';
+import {cumImpl} from './Cum_impl';
 
 export function cumsum(
     args:
@@ -30,55 +28,7 @@ export function cumsum(
   const {inputs, backend, attrs} = args;
   const {x} = inputs;
   const {axis, exclusive, reverse} = attrs;
-
-  const xRank = x.shape.length;
-  const permutation = backend_util.getAxesPermutation([axis], xRank);
-  let permutedX = x;
-  if (permutation != null) {
-    permutedX = transpose({inputs: {x}, backend, attrs: {perm: permutation}});
-  }
-  const permutedAxis = backend_util.getInnerMostAxes(1, xRank)[0];
-
-  if (permutedAxis !== xRank - 1) {
-    throw new Error(
-        `WebGL cumsum shader expects an inner-most axis=${
-            x.shape.length - 1} ` +
-        `but got axis=${axis}`);
-  }
-  const size = permutedX.shape[permutedAxis];
-  let result = identity({inputs: {x: permutedX}, backend});
-  // Use cumsum parallel algorithm, ref:
-  // https://developer.nvidia.com/gpugems/gpugems3/part-vi-gpu-computing/chapter-39-parallel-prefix-sum-scan-cuda
-
-  for (let i = 0; i <= Math.ceil(Math.log2(size)) - 1; i++) {
-    const program = new CumSumProgram(permutedX.shape, false, reverse);
-    const customValues = [[i]];
-    const prevResult = result;
-    result =
-        backend.runWebGLProgram(program, [result], result.dtype, customValues);
-    backend.disposeIntermediateTensorInfo(prevResult);
-  }
-  // For exclusive cumsum, shift the end result in the direction of sum
-  // and add 0 to the front index.
-  if (exclusive) {
-    const program = new CumSumProgram(permutedX.shape, exclusive, reverse);
-    const prevResult = result;
-    result = backend.runWebGLProgram(program, [result], result.dtype);
-    backend.disposeIntermediateTensorInfo(prevResult);
-  }
-
-  if (permutation != null) {
-    const reversePermutation = backend_util.getUndoAxesPermutation(permutation);
-    const reverseTransposedResult = transpose(
-        {inputs: {x: result}, backend, attrs: {perm: reversePermutation}});
-
-    backend.disposeIntermediateTensorInfo(result);
-    backend.disposeIntermediateTensorInfo(permutedX);
-
-    return reverseTransposedResult;
-  }
-
-  return result;
+  return cumImpl(CumOpType.Sum, x, backend, axis, exclusive, reverse);
 }
 
 export const cumsumConfig: KernelConfig = {
